+++ src/aks-preview/azext_aks_preview/custom.py
@@ -104,7 +104,6 @@ from ._consts import CONST_INGRESS_APPGW_WATCH_NAMESPACE
 from ._consts import CONST_SCALE_SET_PRIORITY_REGULAR, CONST_SCALE_SET_PRIORITY_SPOT, CONST_SPOT_EVICTION_POLICY_DELETE
 from ._consts import CONST_CONFCOM_ADDON_NAME, CONST_ACC_SGX_QUOTE_HELPER_ENABLED
 from ._consts import CONST_OPEN_SERVICE_MESH_ADDON_NAME
-from ._consts import CONST_PRIVATE_DNS_ZONE_SYSTEM, CONST_PRIVATE_DNS_ZONE_NONE
 from ._consts import ADDONS
 logger = get_logger(__name__)
 
@@ -1186,9 +1185,6 @@ def aks_create(cmd,     # pylint: disable=too-many-locals,too-many-statements,to
     if private_dns_zone:
         if not enable_private_cluster:
             raise CLIError("Invalid private dns zone for public cluster. It should always be empty for public cluster")
-        # remove following check once we support custom private dns zone
-        if private_dns_zone not in (CONST_PRIVATE_DNS_ZONE_SYSTEM, CONST_PRIVATE_DNS_ZONE_NONE):
-            raise CLIError("Invalid private dns zone for private cluster. Only 'system' or 'none' mode is supported")
         mc.api_server_access_profile.private_dns_zone = private_dns_zone
 
     if uptime_sla: