+++ src/storage-blob-preview/azext_storage_blob_preview/__init__.py
@@ -128,6 +128,15 @@ class StorageArgumentContext(AzArgumentContext):
                    help='Specify a SQL where clause on blob tags to operate only on blobs with a matching value.')
 
     def register_blob_arguments(self):
+        from ._validators import validate_blob_arguments
+        self.extra('blob_name')
+        self.extra('container_name')
+        self.extra('timeout', help='Request timeout in seconds. Applies to each call to the service.', type=int)
+        self.extra('blob_url', help='The full endpoint URL to the Blob, including SAS token and snapshot if used. '
+                   'This could be either the primary endpoint, or the secondary endpoint depending on the current '
+                   '`location_mode`.', validator=validate_blob_arguments)
+
+    def register_lease_blob_arguments(self):
         self.extra('blob_name', required=True)
         self.extra('container_name', required=True)
         self.extra('timeout', help='Request timeout in seconds. Applies to each call to the service.', type=int)