+++ src/connectedk8s/HISTORY.rst
@@ -5,8 +5,9 @@ Release History
 
 0.2.8
 ++++++
-* Added checks for proxy
+* Added checks for proxy and added disable-proxy
 * Updated config dataplane endpoint to support other clouds
+* `az connectedk8s connect`: Added support for kubernetes distro/infra parameters and heuristics
 
 0.2.7
 ++++++